+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     - Symmetrize TSNE affinities matrix, fixes optimization
     - Add numerical stability to SoftPlus activation function
     - Fix Extra Tree leaf node entropy
+    - Fix SVC save/load dropping the class label map
 
 - 2.5.8
     - Optimize squaring operations